third_party/blink/web_tests/platform/linux/external/wpt/fetch/api/credentials/cookies.any.sharedworker-expected.txt

This is a testharness.js-based test.
[FAIL] Include mode: 1 cookie
  assert_equals: Request include cookie(s) expected "a=1" but got "a=1; cspViolationReportCookie2=1"
[FAIL] Include mode: 2 cookies
  assert_equals: Request include cookie(s) expected "b=2; c=3" but got "b=2; c=3; cspViolationReportCookie2=1"
[FAIL] Omit mode: no cookie is stored
  assert_false: Request does not have cookie(s) expected false got true
[FAIL] Same-origin mode: 1 cookie
  assert_equals: Request include cookie(s) expected "a=1" but got "a=1; cspViolationReportCookie2=1"
[FAIL] Same-origin mode: 2 cookies
  assert_equals: Request include cookie(s) expected "b=2; c=3" but got "b=2; c=3; cspViolationReportCookie2=1"
Harness: the test ran to completion.
